git-svn-id: svn://svn.code.sf.net/p/qwt/code/trunk@570 29e17f8d-cf37-471e-a288-72cc03a76eda

+++ qwt/examples/oscilloscope/signalgenerator.h
@@ -16,7 +16,7 @@ public:
 public slots:
     void setAmplitude(double);
     void setFrequency(double);
-    void setSignalInterval(int); // ms
+    void setSignalInterval(double); // ms
 
 protected:
     virtual void run();
@@ -27,5 +27,5 @@ private:
     double d_frequency;
     double d_amplitude;
 
-    int d_signalInterval;
+    double d_signalInterval;
 };
